+++ Motores1/Assets/Niveles/Nivel1.unity
@@ -605,17 +605,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2562824796056252442, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
   m_PrefabInstance: {fileID: 3171604993550939248}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1003800009 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 3137184854273834208, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
-  m_PrefabInstance: {fileID: 3171604993550939248}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1109995460 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2959824634297744251, guid: 73861d7e8fb459744bb8ab5e929ca675, type: 3}
@@ -653,17 +642,6 @@ Transform:
   - {fileID: 350838324}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1171585329 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1517533167814654226, guid: 73861d7e8fb459744bb8ab5e929ca675, type: 3}
-  m_PrefabInstance: {fileID: 9061935639923004302}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1171773147 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2893685329769731614, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
@@ -674,17 +652,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 5348002036545618843, guid: 3ec3d01b7305cda46a7480edab7fd832, type: 3}
   m_PrefabInstance: {fileID: 242620198340287997}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1202756799 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5891842166791996947, guid: 3ec3d01b7305cda46a7480edab7fd832, type: 3}
-  m_PrefabInstance: {fileID: 242620198340287997}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1234516237 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2893685329769731614, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
@@ -700,17 +667,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 8221252675783057818, guid: 40749b845e0f6fb4bb6640c1c616cb68, type: 3}
   m_PrefabInstance: {fileID: 4665395904413657698}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1335988884 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 2700270490962350243, guid: c2fd62a839f1e6f489022679b88609f1, type: 3}
-  m_PrefabInstance: {fileID: 1931106365741692421}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1506199985
 GameObject:
   m_ObjectHideFlags: 0
@@ -921,17 +877,6 @@ MeshFilter:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1612255159}
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
---- !u!114 &1658581427 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 851310811210767248, guid: c16e6e9472b128142a2ee8afaad3df75, type: 3}
-  m_PrefabInstance: {fileID: 5424795497758315251}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1695105671 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2683098557535016777, guid: 2c657307e63acda43b6635ffc31d5021, type: 3}
@@ -1137,17 +1082,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2893685329769731614, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
   m_PrefabInstance: {fileID: 1804979990}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1804979992 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1525311484525297858, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
-  m_PrefabInstance: {fileID: 1804979990}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b883bbd8041653640b7635601aff1ac5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1879391531 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 8181504403212157173, guid: 70bba8adff408434da56f6ad67bbdbf6, type: 3}
@@ -1253,17 +1187,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 4568784160464113180, guid: cd34bf157adfa99428a13e18c098adab, type: 3}
   m_PrefabInstance: {fileID: 6807345324384922694}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &2124586580 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5753854772061812803, guid: cd34bf157adfa99428a13e18c098adab, type: 3}
-  m_PrefabInstance: {fileID: 6807345324384922694}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &2142864668 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 376802974176190589, guid: 20d205cf021ebb74a94af586d2bd2187, type: 3}
@@ -4892,34 +4815,6 @@ PrefabInstance:
       propertyPath: m_AnchoredPosition.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[0]
-      value: 
-      objectReference: {fileID: 2124586580}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[1]
-      value: 
-      objectReference: {fileID: 1202756799}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[2]
-      value: 
-      objectReference: {fileID: 1003800009}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[3]
-      value: 
-      objectReference: {fileID: 1335988884}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[4]
-      value: 
-      objectReference: {fileID: 1658581427}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[5]
-      value: 
-      objectReference: {fileID: 1171585329}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[6]
-      value: 
-      objectReference: {fileID: 1804979992}
     - target: {fileID: 2360276934889744015, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
       propertyPath: _myOwner
       value: 
@@ -4928,6 +4823,10 @@ PrefabInstance:
       propertyPath: _myOwner
       value: 
       objectReference: {fileID: 2089405689}
+    - target: {fileID: 3666585443364198487, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 3999285050153611473, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
       propertyPath: m_AnchorMax.y
       value: 0
