+++ Motores1/Assets/Niveles/Nivel2.unity
@@ -454,17 +454,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 5348002036545618843, guid: 3ec3d01b7305cda46a7480edab7fd832, type: 3}
   m_PrefabInstance: {fileID: 393021952}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &393021954 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5891842166791996947, guid: 3ec3d01b7305cda46a7480edab7fd832, type: 3}
-  m_PrefabInstance: {fileID: 393021952}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &397323698
 GameObject:
   m_ObjectHideFlags: 0
@@ -641,17 +630,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2893685329769731614, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
   m_PrefabInstance: {fileID: 573914852}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &573914854 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1525311484525297858, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
-  m_PrefabInstance: {fileID: 573914852}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b883bbd8041653640b7635601aff1ac5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &692478488 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 1661366912030859346, guid: c2fd62a839f1e6f489022679b88609f1, type: 3}
@@ -719,17 +697,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 4568784160464113180, guid: cd34bf157adfa99428a13e18c098adab, type: 3}
   m_PrefabInstance: {fileID: 698562909}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &698562911 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5753854772061812803, guid: cd34bf157adfa99428a13e18c098adab, type: 3}
-  m_PrefabInstance: {fileID: 698562909}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1001 &847238490
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -849,17 +816,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2959824634297744251, guid: 73861d7e8fb459744bb8ab5e929ca675, type: 3}
   m_PrefabInstance: {fileID: 921264685}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &921264687 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1517533167814654226, guid: 73861d7e8fb459744bb8ab5e929ca675, type: 3}
-  m_PrefabInstance: {fileID: 921264685}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1003800007 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2562824796056252442, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
@@ -1237,17 +1193,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 1661366912030859346, guid: c2fd62a839f1e6f489022679b88609f1, type: 3}
   m_PrefabInstance: {fileID: 1629566388}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1629566390 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 2700270490962350243, guid: c2fd62a839f1e6f489022679b88609f1, type: 3}
-  m_PrefabInstance: {fileID: 1629566388}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1001 &1689995633
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1310,17 +1255,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 9209181727510447363, guid: c16e6e9472b128142a2ee8afaad3df75, type: 3}
   m_PrefabInstance: {fileID: 1689995633}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1689995635 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 851310811210767248, guid: c16e6e9472b128142a2ee8afaad3df75, type: 3}
-  m_PrefabInstance: {fileID: 1689995633}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1695105671 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2683098557535016777, guid: 2c657307e63acda43b6635ffc31d5021, type: 3}
@@ -1526,17 +1460,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2562824796056252442, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
   m_PrefabInstance: {fileID: 1785259370}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1785259372 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 3137184854273834208, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
-  m_PrefabInstance: {fileID: 1785259370}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1879391531 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 8181504403212157173, guid: 70bba8adff408434da56f6ad67bbdbf6, type: 3}
@@ -5286,38 +5209,6 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.size
-      value: 7
-      objectReference: {fileID: 0}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[0]
-      value: 
-      objectReference: {fileID: 698562911}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[1]
-      value: 
-      objectReference: {fileID: 393021954}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[2]
-      value: 
-      objectReference: {fileID: 1785259372}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[3]
-      value: 
-      objectReference: {fileID: 1629566390}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[4]
-      value: 
-      objectReference: {fileID: 1689995635}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[5]
-      value: 
-      objectReference: {fileID: 921264687}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[6]
-      value: 
-      objectReference: {fileID: 573914854}
     - target: {fileID: 2360276934889744015, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
       propertyPath: _myOwner
       value:
