+++ Motores1/Assets/Niveles/Nivel3.unity
@@ -287,17 +287,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 1661366912030859346, guid: c2fd62a839f1e6f489022679b88609f1, type: 3}
   m_PrefabInstance: {fileID: 227902211}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &227902213 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 2700270490962350243, guid: c2fd62a839f1e6f489022679b88609f1, type: 3}
-  m_PrefabInstance: {fileID: 227902211}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &340744359 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 6481887279343516005, guid: 986ecf699c0c1a545b2ae771ca2edb07, type: 3}
@@ -612,17 +601,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2959824634297744251, guid: 73861d7e8fb459744bb8ab5e929ca675, type: 3}
   m_PrefabInstance: {fileID: 403643568}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &403643570 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1517533167814654226, guid: 73861d7e8fb459744bb8ab5e929ca675, type: 3}
-  m_PrefabInstance: {fileID: 403643568}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1001 &409063819
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -747,17 +725,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2893685329769731614, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
   m_PrefabInstance: {fileID: 467383577}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &467383579 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1525311484525297858, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
-  m_PrefabInstance: {fileID: 467383577}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b883bbd8041653640b7635601aff1ac5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &546238882 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 902283556761338133, guid: f548dfc8703c3a542bbdcc3749357153, type: 3}
@@ -830,17 +797,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 5348002036545618843, guid: 3ec3d01b7305cda46a7480edab7fd832, type: 3}
   m_PrefabInstance: {fileID: 724235149}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &724235151 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5891842166791996947, guid: 3ec3d01b7305cda46a7480edab7fd832, type: 3}
-  m_PrefabInstance: {fileID: 724235149}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1001 &847238490
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -960,17 +916,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 4568784160464113180, guid: cd34bf157adfa99428a13e18c098adab, type: 3}
   m_PrefabInstance: {fileID: 893542306}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &893542308 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5753854772061812803, guid: cd34bf157adfa99428a13e18c098adab, type: 3}
-  m_PrefabInstance: {fileID: 893542306}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1003800007 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2562824796056252442, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
@@ -1090,17 +1035,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2562824796056252442, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
   m_PrefabInstance: {fileID: 1268813013}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1268813015 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 3137184854273834208, guid: dc212d166e09b0244b58317835b3b8d8, type: 3}
-  m_PrefabInstance: {fileID: 1268813013}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bbcbb313f99e79d48ab7c6563852a87d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1288387585 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 2893685329769731614, guid: 13b98699c4fb59e4e820aa508f927126, type: 3}
@@ -1173,17 +1107,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 9209181727510447363, guid: c16e6e9472b128142a2ee8afaad3df75, type: 3}
   m_PrefabInstance: {fileID: 1305114517}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1305114519 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 851310811210767248, guid: c16e6e9472b128142a2ee8afaad3df75, type: 3}
-  m_PrefabInstance: {fileID: 1305114517}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5beb07bd28b31243b130b9966e1584a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1506199985
 GameObject:
   m_ObjectHideFlags: 0
@@ -5290,38 +5213,6 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.size
-      value: 7
-      objectReference: {fileID: 0}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[0]
-      value: 
-      objectReference: {fileID: 893542308}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[1]
-      value: 
-      objectReference: {fileID: 724235151}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[2]
-      value: 
-      objectReference: {fileID: 1268813015}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[3]
-      value: 
-      objectReference: {fileID: 227902213}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[4]
-      value: 
-      objectReference: {fileID: 1305114519}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[5]
-      value: 
-      objectReference: {fileID: 403643570}
-    - target: {fileID: 1901919104818204617, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
-      propertyPath: shopItems.Array.data[6]
-      value: 
-      objectReference: {fileID: 467383579}
     - target: {fileID: 2360276934889744015, guid: 56edde2eed1965149b5ca132b9e930fe, type: 3}
       propertyPath: _myOwner
       value:
