+++ Motores1/Assets/Scripts/ItemsComprables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d744f26f351288e4b90d8cbe6a98eae1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
