+++ Motores1/Assets/Scripts/ItemsComprables/ChestComprable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9aa32b2f906d92d47ab8e3c1dd62f44a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
