+++ Motores1/Assets/Scripts/ItemsComprables/GreatSwordComprable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70f6a0a4cf47bf9428e6d63d76c4ff45
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
