+++ Motores1/Assets/Scripts/ItemsComprables/HelmetComprable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 751ec22abbb3fcd4386f55f2300fd17c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
