+++ Motores1/Assets/Scripts/ItemsComprables/ItemComprable.cs.meta
