+++ Motores1/Assets/Scripts/ItemsComprables/LegsComprable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1edb52a996449704f9d42ea5b64b02d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
