+++ Motores1/Assets/Scripts/ItemsComprables/PotionComprable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7a86ff1021fa1704585d4ebad0d45c1d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
