+++ Motores1/Assets/Scripts/ItemsComprables/ShielComprable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76ef4d58e6f96a9479e5ed8931f8bfa9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
