+++ Motores1/Assets/Scripts/ItemsComprables/SwordComprable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1980be6a316a9cb40a620a7e6243dad9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
