+++ Motores1/Assets/Scripts/ObjetosInteractuables/Weapon.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 using static UnityEditor.Experimental.GraphView.GraphView;
 using static Weapon;
 
-public class Weapon : ItemComprable, IPickeable
+public class Weapon : MonoBehaviour, IPickeable
 {
     public WeaponType weaponType; 
 
@@ -16,76 +16,76 @@ public class Weapon : ItemComprable, IPickeable
         StartCoroutine(ActiveCosa(5f));
     }
 
-    public override void BuyItem()
-    {
-        switch (weaponType)
-        {
-            case WeaponType.Sword:
-                BuySword();
-                break;
-            case WeaponType.GreatSword:
-                BuyGreatSword();
-                break;
-            case WeaponType.SwordAndShield:
-                BuySandShield();
-                break;
-        }
-    }
+    //public override void BuyItem()
+    //{
+    //    switch (weaponType)
+    //    {
+    //        case WeaponType.Sword:
+    //            BuySword();
+    //            break;
+    //        case WeaponType.GreatSword:
+    //            BuyGreatSword();
+    //            break;
+    //        case WeaponType.SwordAndShield:
+    //            BuySandShield();
+    //            break;
+    //    }
+    //}
 
-    public void BuySword()
-    {
-        if (GameManager.Instance.Player.playerWeapon == Weapon.WeaponType.Sword)
-        {
-            Debug.Log("<color=red> Arma ya equipada </color>");
-            return;
-        }
-        if (!GameManager.Instance.Player.PayAmount(price))
-        {
-            Debug.Log("<color=yellow>Dinero insuficiente</color>");
-            return;
-        }
-        GameManager.Instance.Player.SetWeapon(Weapon.WeaponType.Sword);
+    //public void BuySword()
+    //{
+    //    if (GameManager.Instance.Player.playerWeapon == Weapon.WeaponType.Sword)
+    //    {
+    //        Debug.Log("<color=red> Arma ya equipada </color>");
+    //        return;
+    //    }
+    //    if (!GameManager.Instance.Player.PayAmount(price))
+    //    {
+    //        Debug.Log("<color=yellow>Dinero insuficiente</color>");
+    //        return;
+    //    }
+    //    GameManager.Instance.Player.SetWeapon(Weapon.WeaponType.Sword);
 
-        //_weaponType = Weapon.WeaponType.Sword;
+    //    //_weaponType = Weapon.WeaponType.Sword;
 
-    }
-    public void BuyGreatSword()
-    {
-        if (GameManager.Instance.Player.playerWeapon == Weapon.WeaponType.GreatSword)
-        {
-            Debug.Log("<color=red> Arma ya equipada </color>");
-            return;
-        }
-        if (!GameManager.Instance.Player.PayAmount(price))
-        {
-            Debug.Log("<color=yellow>Dinero insuficiente</color>");
-            return;
-        }
-        GameManager.Instance.Player.SetWeapon(Weapon.WeaponType.GreatSword);
+    //}
+    //public void BuyGreatSword()
+    //{
+    //    if (GameManager.Instance.Player.playerWeapon == Weapon.WeaponType.GreatSword)
+    //    {
+    //        Debug.Log("<color=red> Arma ya equipada </color>");
+    //        return;
+    //    }
+    //    if (!GameManager.Instance.Player.PayAmount(price))
+    //    {
+    //        Debug.Log("<color=yellow>Dinero insuficiente</color>");
+    //        return;
+    //    }
+    //    GameManager.Instance.Player.SetWeapon(Weapon.WeaponType.GreatSword);
 
-        //_weaponType = Weapon.WeaponType.GreatSword;
-        //_playerMoney -= weaponCost;
+    //    //_weaponType = Weapon.WeaponType.GreatSword;
+    //    //_playerMoney -= weaponCost;
 
 
-    }
-    public void BuySandShield()
-    {
-        if (GameManager.Instance.Player.playerWeapon == Weapon.WeaponType.SwordAndShield)
-        {
-            Debug.Log("<color=red> Arma ya equipada </color>");
-            return;
-        }
-        if (!GameManager.Instance.Player.PayAmount(price))
-        {
-            Debug.Log("<color=yellow>Dinero insuficiente</color>");
-            return;
-        }
-        GameManager.Instance.Player.SetWeapon(Weapon.WeaponType.SwordAndShield);
+    //}
+    //public void BuySandShield()
+    //{
+    //    if (GameManager.Instance.Player.playerWeapon == Weapon.WeaponType.SwordAndShield)
+    //    {
+    //        Debug.Log("<color=red> Arma ya equipada </color>");
+    //        return;
+    //    }
+    //    if (!GameManager.Instance.Player.PayAmount(price))
+    //    {
+    //        Debug.Log("<color=yellow>Dinero insuficiente</color>");
+    //        return;
+    //    }
+    //    GameManager.Instance.Player.SetWeapon(Weapon.WeaponType.SwordAndShield);
 
-        //_weaponType = Weapon.WeaponType.SwordAndShield;
-        //_playerMoney -= weaponCost;
+    //    //_weaponType = Weapon.WeaponType.SwordAndShield;
+    //    //_playerMoney -= weaponCost;
 
-    }
+    //}
 
 
     //momentaneo para testeo
